@@ -37,8 +37,6 @@ dependencies:
 #  rxdart: ^0.27.3
   equatable: ^2.0.3
   json_annotation: ^4.4.0
-  #firebase_core: ^1.12.0
-  #firebase_auth: ^1.1.1
   shared_preferences: ^2.0.14
   google_maps_cluster_manager: ^3.0.0+1
   carousel_slider: ^4.1.1
@@ -46,9 +44,26 @@ dependencies:
   dio: ^4.0.6
   webview_flutter: ^3.0.4
   shaky_animated_listview: ^0.0.5
-
+  #package_info_plus: ^1.4.2
+  bloc_concurrency: ^0.2.0
+#  firebase_auth: ^3.4.2
+#  firebase_core: ^1.19.2
+  google_sign_in: ^5.4.0
+  loading_animation_widget: ^1.2.0+2
+  test: ^1.21.4
+  flutter_easyloading: ^3.0.5
 
   network_manager: ^0.1.0
+  dio_cookie_manager: ^2.0.0
+  network_info_plus: ^2.3.0
+  flutter_secure_storage: ^6.0.0
+  jwt_decoder: ^2.0.1
+  get_it: ^7.2.0
+
+#  path_provider: ^0.4.1
+#  pspdfkit_flutter:
+#    git:
+#      url: git://github.com/PSPDFKit/pspdfkit-flutter.git
 #  latlong2: ^0.8.1
 
   #flutter_svg: ^1.0.2
@@ -59,10 +74,18 @@ dependencies:
   # Use with the CupertinoIcons class for iOS style icons.
   cupertino_icons: ^1.0.2
   flutter_svg: ^1.0.2
-  google_maps_flutter: ^2.1.1
+  google_maps_flutter: ^2.2.0
   flutter_google_places: ^0.3.0
   google_maps_webservice: ^0.0.19
   latlong2: ^0.8.1
+  custom_refresh_indicator: ^1.2.1
+  geolocator: ^9.0.1
+  marquee: ^2.2.3
+  flip_widget: ^0.1.11
+  better_page_turn: ^0.0.6
+  flutter_spinkit: ^5.1.0
+  image_size_getter: ^2.1.2
+  page_curl: ^0.0.1
 
   expandable: ^5.0.1
 
@@ -100,16 +123,16 @@ flutter:
 
   # To add assets to your application, add an assets section, like this:
   # assets:
-  #   - images/a_dot_burr.jpeg
+  #   - images/a_dot_burr.jpeg®
   #   - images/a_dot_ham.jpeg
 
   assets:
     - assets/images/
-    - assets/images/pins.svg
-    - assets/images/pins-2.svg
-    - assets/images/pins-3.svg
+    - assets/images/2.0x/
 
-  # An image asset can refer to one or more resolution-specific "variants", see
+    - assets/images/pins/
+
+  # An image asset can refer to one or more resolution-specific "variants", seeq
   # https://flutter.dev/assets-and-images/#resolution-aware.
 
   # For details regarding adding assets from package dependencies, see
