@@ -4,7 +4,7 @@ part of 'auth_bloc.dart';
 abstract class AuthState extends Equatable {}
 
 // When the user presses the signin or signup button the state is changed to loading first and then to Authenticated.
-class Loading extends AuthState {
+class LoadingAuth extends AuthState {
   @override
   List<Object?> get props => [];
 }
