@@ -1,7 +1,14 @@
+import 'dart:convert';
+
 import 'dart:typed_data';
 
 import 'package:bloc/bloc.dart';
+import 'package:dio/adapter.dart';
+import 'package:dio/dio.dart';
 import 'package:equatable/equatable.dart';
+import 'package:sharemagazines_flutter/src/resources/magazine_repository.dart';
+
+import '../constants.dart';
 
 part 'reader_event.dart';
 part 'reader_state.dart';
@@ -9,69 +16,90 @@ part 'reader_state.dart';
 // enum NavbarItems { Home, Menu, Map, Account }
 
 class ReaderBloc extends Bloc<ReaderEvent, ReaderState> {
+  final MagazineRepository magazineRepository;
   // final MagazineRepository magazineRepository;
   // late ReaderState currentState;
-  ReaderBloc() : super(Initialized()) {
-    on<Initialize>((event, emit) async {
-      emit(Initialized());
-      // try {
-      //   print("emission-1");
-      //   // final magazinePublishedGetLastWithLimitdata = await magazineRepository
-      //   //     .magazinePublishedGetLastWithLimit(id_hotspot: '0');
-      //   // // print(magazinePublishedGetLastWithLimit!.response);
-      //   print("blocc");
-      //   // currentState = GoToHome(
-      //   //     magazinePublishedGetLastWithLimit:
-      //   //         magazinePublishedGetLastWithLimitdata!);
-      //   final getimage =
-      //       await magazineRepository.Getimage(page: '0', id_mag_pub: '190360');
-      //   print(getimage);
-      //   print("blocc");
-      //   emit(GoToHome(magazinePublishedGetLastWithLimitdata, getimage));
-      // } catch (e) {
-      //   print("noemission");
-      //   emit(NavbarError(e.toString()));
-      // }
-    });
+  late List<Uint8List> image = [];
+  late Uint8List image123;
+  late List<Future<Uint8List>> futureFuncAllPages = List.empty(growable: true);
+  // final dio = Dio();
+  // late List<Future<Uint8List>> futureFunc = List.empty(growable: true);
+  ReaderBloc({required this.magazineRepository}) : super(Initialized()) {
+    // on<Initialize>((event, emit) async {
+    //   emit(Initialized());
+    //   // try {
+    //   //   print("emission-1");
+    //   //   // final magazinePublishedGetLastWithLimitdata = await magazineRepository
+    //   //   //     .magazinePublishedGetLastWithLimit(id_hotspot: '0');
+    //   //   // // print(magazinePublishedGetLastWithLimit!.response);
+    //   //   print("blocc");
+    //   //   // currentState = GoToHome(
+    //   //   //     magazinePublishedGetLastWithLimit:
+    //   //   //         magazinePublishedGetLastWithLimitdata!);
+    //   //   final getimage =
+    //   //       await magazineRepository.Getimage(page: '0', id_mag_pub: '190360');
+    //   //   print(getimage);
+    //   //   print("blocc");
+    //   //   emit(GoToHome(magazinePublishedGetLastWithLimitdata, getimage));
+    //   // } catch (e) {
+    //   //   print("noemission");
+    //   //   emit(NavbarError(e.toString()));
+    //   // }
+    // });
 
     on<OpenReader>((event, emit) async {
-      emit(ReaderOpened());
-      // print("bloccc");
-      // try {
-      //   final magazinePublishedGetLastWithLimit = await magazineRepository
-      //       .magazinePublishedGetLastWithLimit(id_hotspot: '0');
-      //   emit(GoToHome(magazinePublishedGetLastWithLimit!));
-      // } catch (e) {
-      //   emit(NavbarError(e.toString()));
+      // ReaderOpened(futureFunc);
+      print("OpenReader");
+      print(" OpenReader ${event.idMagazinePublication}");
+      // futureFuncAllPages = futureFunc;
+      // dio.options.baseUrl = ApiConstants.baseUrl;
+      // dio.interceptors.add(
+      //   InterceptorsWrapper(
+      //     onRequest: (
+      //       RequestOptions requestOptions,
+      //       RequestInterceptorHandler handler,
+      //     ) {
+      //       print(requestOptions.uri);
+      //       Future.delayed(Duration(seconds: 5), () {
+      //         handler.next(requestOptions);
+      //       });
+      //     },
+      //   ),
+      // );
+      // dio.httpClientAdapter = DefaultHttpClientAdapter()..onHttpClientCreate = (httpClient) => httpClient..maxConnectionsPerHost = 20;
+      // for (var i = 0; i < int.parse(event.pageNo) + 1; i++) {
+      //   // id_mag_pub: '51');
+      //   // imagebytes.add(image123);
+      //
+      //   image123 = await magazineRepository.GetPage(page: i.toString(), id_mag_pub: event.idMagazinePublication, client: dio);
+      //   // id_mag_pub: '51');
+      //   image.add(image123);
+      //
+      //   // var image1 = File(image).writeAsBytes(Uint8list);
+      //
+      //   // print("bloccc");
+      //   // // print(image1);
+      //   emit(ReaderOpened(image));
       // }
-      // });
-      // on<Menu>((event, emit) async {
-      //   emit(GoToMenu());
-      // });
-      // on<Map>((event, emit) async {
-      //   emit(GoToMap());
-      // });
-      // on<AccountEvent>((event, emit) async {
-      //   emit(GoToAccount());
-      // });
-      // on<Location>((event, emit) async {
-      //   emit(GoToLocation());
+      //
+      // emit(ReaderOpened(image));
+      try {
+        for (var i = 0; i < int.parse(event.pageNo); i++) {
+          futureFuncAllPages.add(magazineRepository.GetPagesforReader(page: i.toString(), id_mag_pub: event.idMagazinePublication));
+        }
+      } finally {
+        emit(ReaderOpened(futureFuncAllPages));
+      }
     });
+    // print(image);
+    // print("bloccc");
+    // try {
+    //   final magazinePublishedGetLastWithLimit = await magazineRepository
+    //       .magazinePublishedGetLastWithLimit(id_hotspot: '0');
+    //   emit(GoToHome(magazinePublishedGetLastWithLimit!));
+    // } catch (e) {
+    //   emit(NavbarError(e.toString()));
+    // }
+    // });
   }
-// void getNavBarItem(NavbarItems navbarItem) {
-//   switch (navbarItem) {
-//     case NavbarItems.Home:
-//       emit(NavbarState(NavbarItems.Home, 0));
-//       break;
-//     case NavbarItems.Menu:
-//       emit(NavbarState(NavbarItems.Menu, 1));
-//       break;
-//     case NavbarItems.Map:
-//       emit(NavbarState(NavbarItems.Map, 2));
-//       break;
-//     case NavbarItems.Account:
-//       emit(NavbarState(NavbarItems.Account, 3));
-//       break;
-//   }
-// }
 }
