@@ -1,3 +1,4 @@
+import 'dart:async';
 import 'dart:developer';
 import 'dart:io';
 
@@ -6,87 +7,158 @@ import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter/widgets.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
+import 'package:flutter_easyloading/flutter_easyloading.dart';
 import 'package:flutter_svg/svg.dart';
 import 'package:sharemagazines_flutter/src/blocs/auth_bloc.dart';
 import 'package:sharemagazines_flutter/src/blocs/map_bloc.dart';
 import 'package:sharemagazines_flutter/src/blocs/navbar_bloc.dart';
+import 'package:sharemagazines_flutter/src/blocs/splash_bloc.dart';
 import 'package:sharemagazines_flutter/src/presentation/pages/homepage.dart';
+import 'package:sharemagazines_flutter/src/presentation/pages/menupage.dart';
+import 'package:sharemagazines_flutter/src/presentation/pages/searchpage.dart';
 import 'package:sharemagazines_flutter/src/presentation/pages/startpage.dart';
 import 'package:sharemagazines_flutter/src/presentation/pages/mappage.dart';
 import 'package:sharemagazines_flutter/src/presentation/widgets/account.dart';
 import 'package:sharemagazines_flutter/src/presentation/widgets/body_painter.dart';
 import 'package:sharemagazines_flutter/src/presentation/widgets/bottomAppBar.dart';
-import 'package:sharemagazines_flutter/src/presentation/widgets/locationAppbar.dart';
+import 'package:sharemagazines_flutter/src/presentation/widgets/SlidingAppBar.dart';
+import 'package:sharemagazines_flutter/src/presentation/widgets/marquee.dart';
 import 'package:sharemagazines_flutter/src/presentation/widgets/navbar_painter.dart';
 import 'package:sharemagazines_flutter/src/resources/hotspot_repository.dart';
 import 'package:sharemagazines_flutter/src/resources/magazine_repository.dart';
 
-class MainPageState extends StatelessWidget {
-  const MainPageState({Key? key}) : super(key: key);
-
-  @override
-  Widget build(BuildContext context) {
-    return BlocProvider(
-        create: (context) => NavbarBloc(
-            magazineRepository:
-                RepositoryProvider.of<MagazineRepository>(context)),
-        child: MainPage());
-  }
-}
+import '../../blocs/search_bloc.dart' as searchBloc;
+import '../../resources/location_repository.dart';
 
 class MainPage extends StatefulWidget {
   MainPage({Key? key}) : super(key: key);
-  int num = 0;
-
-  // final NavbarBloc _navbarBloc;
 
   @override
   State<MainPage> createState() => _MainPageState();
 }
 
-class _MainPageState extends State<MainPage> {
+class _MainPageState extends State<MainPage> with SingleTickerProviderStateMixin, AutomaticKeepAliveClientMixin {
   // NavbarBloc _navbarBloc = BlocProvider.of<NavbarBloc>(context);
+
+  late final AnimationController _controller;
   int selectedIndex = 0;
+  late PageController _pageController;
+  var showSearchPage = false;
+
+  //easyloading
+  Timer? _timer;
+  late double _progress;
+  // _MainPageState(){
+  //
+  // }
+  late List<Widget> _pages = [
+    // HomePageState(navbarBloc: BlocProvider.of<NavbarBloc>(context)),
+    HomePage(),
+    MenuPage(),
+    // MapsBlocWidget(),
+    // Maps(_progress, _timer),
+    Maps(),
+    Account(),
+  ];
   GlobalKey _keyheight = GlobalKey();
   final controller = ScrollController();
   late OverlayEntry sticky;
   GlobalKey stickyKey = GlobalKey();
 
   @override
+  bool get wantKeepAlive => true;
+
+  @override
   void initState() {
     super.initState();
     print("init");
-    BlocProvider.of<NavbarBloc>(context).add(Home());
+    EasyLoading.addStatusCallback((status) {
+      print('EasyLoading Status $status');
+      if (status == EasyLoadingStatus.dismiss) {
+        _timer?.cancel();
+      }
+    });
+    _progress = 0;
+    // EasyLoading.showSuccess('Use in initState');
+    BlocProvider.of<NavbarBloc>(context).add(Initialize123(_timer, _progress));
+
+    _controller = AnimationController(
+      vsync: this,
+      duration: Duration(milliseconds: 400),
+    );
+
+    _pageController = PageController(initialPage: selectedIndex);
+
     // _navbarBloc = NavbarBloc();
+    // _progress = 0;
+    // _timer?.cancel();
+    // _timer = Timer.periodic(const Duration(milliseconds: 100), (Timer timer) {
+    // _timer = Timer.periodic(const Duration(milliseconds: 100), (Timer timer) {
+    //   EasyLoading.showProgress(
+    //     _progress,
+    //     status: '${(_progress * 100).toStringAsFixed(0)}%',
+    //     maskType: EasyLoadingMaskType.black,
+    //   );
+    //   _progress += 0.03;
+    //
+    //   if (_progress >= 1) {
+    //     _timer?.cancel();
+    //     EasyLoading.dismiss();
+    //   }
+    // });
+    // timer = Timer.periodic(const Duration(milliseconds: 100), (Timer timer) {
+    //   EasyLoading.showProgress(_progress, status: '${(_progress * 100).toStringAsFixed(0)}%');
+    //   _progress += 0.03;
+    //
+    //   if (_progress >= 1) {
+    //     _timer?.cancel();
+    //     EasyLoading.dismiss();
+    //   }
+    // });
   }
 
   @override
   void dispose() {
-    // _navbarBloc.dispose();
-    super.dispose();
+    // _pageController.dispose();
+    // _controller.dispose();
+    // super.dispose();
   }
 
   void onClicked(int index) {
+    print("mainpage onclicked");
+    if (showSearchPage == false) {
+      setState(() {
+        _pageController.animateToPage(index, duration: Duration(milliseconds: 300), curve: Curves.ease);
+      });
+    }
+    print(index);
     if (index == 0) {
-      print("emit");
+      print("home page bloc add");
       BlocProvider.of<NavbarBloc>(context).add(
         Home(),
       );
     } else if (index == 1) {
+      print("menu page bloc add");
       BlocProvider.of<NavbarBloc>(context).add(
         Menu(),
       );
     } else if (index == 2) {
-      print("aads");
+      print("Mappage bloc add");
       BlocProvider.of<NavbarBloc>(context).add(
         Map(),
       );
     } else if (index == 3) {
-      print("emit");
+      print("Account page bloc add");
       BlocProvider.of<NavbarBloc>(context).add(
         AccountEvent(),
       );
     }
+    // else if (index == 4) {
+    //   print("Account page bloc add");
+    //   BlocProvider.of<NavbarBloc>(context).add(
+    //     Search(),
+    //   );
+    // }
   }
 
   // @override
@@ -112,8 +184,10 @@ class _MainPageState extends State<MainPage> {
   // }
   @override
   Widget build(BuildContext context) {
+    super.build(context);
+
     return BlocBuilder<NavbarBloc, NavbarState>(builder: (context, state) {
-      int currentIndex = state is GoToHome
+      int currentIndex = state is GoToHome //|| state is HomeLoaded
           ? 0
           : state is GoToMenu
               ? 1
@@ -121,87 +195,144 @@ class _MainPageState extends State<MainPage> {
                   ? 2
                   : state is GoToAccount
                       ? 3
+                      // : state is GoToSearch
+                      //     ? 4
                       : 0;
+      print("Mainpage blocbuilder");
+      // BlocProvider.of<searchBloc.SearchBloc>(context).add(searchBloc.Initialize(context));
+      // int currentIndex = 0;
+      // print(state.access_location.data!.isNotEmpty);
       return Stack(
         children: <Widget>[
-          new SvgPicture.asset(
-            "assets/images/background_webreader.svg",
-            fit: BoxFit.fill,
-            allowDrawingOutsideViewBox: true,
-          ),
-          Scaffold(
-            extendBodyBehindAppBar: true,
-            // extendBody: true,
-            appBar: currentIndex == 0 || currentIndex == 1
-                ? PreferredSize(
-                    preferredSize: const Size(0, 81),
+          // new SvgPicture.asset(
+          //   "assets/images/background_webreader.svg",
+          //   fit: BoxFit.fill,
+          //   allowDrawingOutsideViewBox: true,
+          // ),
+          DecoratedBox(
+            decoration: BoxDecoration(
+              image: DecorationImage(image: AssetImage("assets/images/Background.png"), fit: BoxFit.cover),
+            ),
+            child: Center(
+                child: Scaffold(
+              extendBodyBehindAppBar: currentIndex == 0 || currentIndex == 1 ? true : false,
+              // extendBody: true,
+              // appBar: currentIndex == 0 || currentIndex == 1
+              //     ? PreferredSize(
+              //         preferredSize: const Size(0, 56),
+              appBar: SlidingAppBar(
+                controller: _controller,
+                visible: currentIndex == 0 || currentIndex == 1 ? true : false,
+                child: AppBar(
+                  backgroundColor: Colors.transparent,
+                  elevation: 0,
+                  // toolbarOpacity: currentIndex == 0 || currentIndex == 1 ? 0 : 0,
+                  toolbarHeight: (currentIndex == 0 || currentIndex == 1 && !showSearchPage) ? 100 : 0,
+                  flexibleSpace: SafeArea(
                     child: Container(
-                      // decoration: BoxDecoration(
-                      //     border: Border.all(color: Colors.blueGrey),
-                      //     borderRadius: BorderRadius.all(Radius.circular(20))),
-                      margin: EdgeInsets.only(
-                          left: 10, right: 10, top: 40, bottom: 10),
-                      // padding: EdgeInsets.all(10),
-                      padding: EdgeInsets.only(top: 20, bottom: 0),
+                      margin: EdgeInsets.fromLTRB(10, 0, 10, 0),
                       child: ExpandableNotifier(
                         child: ScrollOnExpand(
-                          child: ExpansionTile(
-                            // onExpansionChanged: (bool expanding) =>
-                            //     _onExpansion(expanding),
-                            leading: Icon(
-                              Icons.circle,
-                              color: Colors.white,
-                              size: 60,
-                            ),
-                            trailing: GestureDetector(
-                              // onTap: ,
-                              child: Icon(
-                                Icons.search,
-                                color: Colors.white,
-                                size: 30,
+                          child: Row(
+                            mainAxisAlignment: MainAxisAlignment.spaceBetween,
+                            children: <Widget>[
+                              Padding(
+                                padding: const EdgeInsets.all(8.0),
+                                child: Icon(
+                                  Icons.circle,
+                                  color: Colors.white,
+                                  size: 80,
+                                ),
                               ),
-                            ),
-                            title: Row(
-                              children: [
-                                Column(
+                              Expanded(
+                                child: Column(
                                   crossAxisAlignment: CrossAxisAlignment.start,
                                   mainAxisAlignment: MainAxisAlignment.center,
                                   children: [
-                                    Container(
+                                    MarqueeWidget(
                                       child: Text(
-                                        "asd",
-                                        // state.magazinePublishedGetLastWithLimit
-                                        //     .response!.first.name!,
-                                        style: TextStyle(
-                                            fontSize: 18, color: Colors.white),
-                                        // textAlign: TextAlign.left,
+                                        // state.access_location.data!.isNotEmpty ==
+                                        //         true
+                                        //     ? "sdf"
+                                        //     : "sdf",
+                                        // "Weser kurier",
+                                        state is GoToHome
+                                            ? state.location?.data[0].nameApp ?? "Please pull down"
+                                            : state is GoToMenu
+                                                ? state.location?.data![0].nameApp ?? "Please pull down"
+                                                : "${state}",
+// state.magazinePublishedGetLastWithLimit
+//     .response!.first.name!,
+//                                         overflow: TextOverflow.e,
+//                                         softWrap: true,
+//                                         maxLines: 1,
+                                        style: TextStyle(fontFamily: 'Raleway', fontSize: 22, color: Colors.white),
+// textAlign: TextAlign.left,
                                       ),
                                     ),
-                                    Container(
-                                      child: Text(
-                                        "Infos zu deiner Location",
-                                        style: TextStyle(
-                                            fontSize: 13,
-                                            color: Colors.white,
-                                            fontWeight: FontWeight.w200),
-                                        // textAlign: TextAlign.right,
-                                      ),
+                                    Text(
+                                      "Infos zu deiner Location",
+                                      // overflow: TextOverflow.fade,
+                                      // softWrap: true,
+                                      style: TextStyle(fontSize: 13, color: Colors.white, fontWeight: FontWeight.w200),
+// textAlign: TextAlign.right,
                                     ),
                                   ],
                                 ),
-                              ],
-                            ),
-                            children: <Widget>[
-                              SingleChildScrollView(
-                                child: Container(
-                                  height:
-                                      MediaQuery.of(context).size.height * 0.8,
-                                  width: MediaQuery.of(context).size.width,
-                                  child: Text(
-                                    "data",
-                                    style: TextStyle(
-                                        fontSize: 20, color: Colors.white),
-                                    textAlign: TextAlign.center,
+                              ),
+                              // Spacer(),
+                              Padding(
+                                padding: EdgeInsets.only(left: 10),
+                                child: InkWell(
+                                  onTap: () => {
+                                    if (mounted) {setState(() => showSearchPage = true)},
+                                    // setState(() {
+                                    //   showSearchPage = true;
+                                    // }),
+                                    print("before searchpage state $state"),
+                                    Navigator.push(
+                                      context,
+                                      PageRouteBuilder(
+                                        // transitionDuration:
+                                        // Duration(seconds: 2),
+
+                                        pageBuilder: (_, __, ___) {
+                                          // return StartSearch();
+
+                                          return SearchPage();
+                                        },
+                                        // maintainState: true,
+
+                                        // transitionDuration: Duration(milliseconds: 1000),
+                                        // transitionsBuilder: (context, animation, anotherAnimation, child) {
+                                        //   // animation = CurvedAnimation(curve: curveList[index], parent: animation);
+                                        //   return ScaleTransition(
+                                        //     scale: animation,
+                                        //     alignment: Alignment.topRight,
+                                        //     child: child,
+                                        //   );
+                                        // }
+                                      ),
+                                    ).then((_) {
+                                      // if (mounted) {
+                                      print("after searchpage state $state");
+
+                                      setState(() {
+                                        showSearchPage = false;
+                                      });
+                                      // }
+                                    })
+                                    // ).then((_) => setState(() {
+                                    //       showSearchPage = false;
+                                    //     }))
+                                  },
+                                  child: Hero(
+                                    tag: "search button",
+                                    child: Icon(
+                                      Icons.search,
+                                      color: Colors.white,
+                                      size: 40,
+                                    ),
                                   ),
                                 ),
                               ),
@@ -210,79 +341,27 @@ class _MainPageState extends State<MainPage> {
                         ),
                       ),
                     ),
-                    // child: AppBar(
-                    //   titleSpacing: 0,
-                    //   leadingWidth: 80,
-                    //   automaticallyImplyLeading: false,
-                    //   backgroundColor: Colors.transparent,
-                    //   leading: Icon(
-                    //     Icons.circle,
-                    //     color: Colors.white,
-                    //     size: 60,
-                    //   ),
-                    //   title: Row(
-                    //     children: [
-                    //       Column(
-                    //         crossAxisAlignment: CrossAxisAlignment.start,
-                    //         mainAxisAlignment: MainAxisAlignment.center,
-                    //         children: [
-                    //           Container(
-                    //             child: Text(
-                    //               "asd",
-                    //               // state.magazinePublishedGetLastWithLimit.response!.first
-                    //               //     .name!,
-                    //               style: TextStyle(fontSize: 18, color: Colors.white),
-                    //               // textAlign: TextAlign.left,
-                    //             ),
-                    //           ),
-                    //           Container(
-                    //             child: Text(
-                    //               "Infos zu deiner Location",
-                    //               style: TextStyle(
-                    //                   fontSize: 13,
-                    //                   color: Colors.white,
-                    //                   fontWeight: FontWeight.w200),
-                    //               // textAlign: TextAlign.right,
-                    //             ),
-                    //           ),
-                    //         ],
-                    //       ),
-                    //     ],
-                    //   ),
-                    // ),
-                  )
-                : null,
-            backgroundColor: Colors.transparent,
-            body: Stack(
-              children: [
-                ClipPath(
-                  clipper: BodyPainter(currentIndex, context, _keyheight),
-                  // clipBehavior: Clip.hardEdge,
-                  child: Container(
-                    child: BlocBuilder<NavbarBloc, NavbarState>(
-                        builder: (context, state) {
-                      if (state is GoToHome || state is GoToMenu) {
-                        // return LocationAppbar();
-                        return SafeArea(child: HomePageState());
-                        // } else if (state is GoToMenu) {
-                        // return MenuPage();
-                        return Container();
-                      } else if (state is GoToMap) {
-                        return MapsBlocWidget();
-                      } else if (state is GoToAccount) {
-                        return Account();
-                      }
-                      ;
-                      return Container();
-                    }),
                   ),
                 ),
-                ClipPath(
-                  key: _keyheight,
-                  clipper: NavBarClipper(currentIndex),
-                  child: Container(
+              ),
+
+              backgroundColor: Colors.transparent,
+              body: Stack(
+                children: [
+                  ClipPath(
+                    clipper: BodyPainter(currentIndex, context, _keyheight),
+                    // clipBehavior: Clip.hardEdge,
+                    child: PageView(
+                      controller: _pageController,
+                      physics: NeverScrollableScrollPhysics(),
+                      // physics: AlwaysScrollableScrollPhysics(),
+                      children: _pages,
+                    ),
+                  ),
+                  // currentIndex != 4
+                  Container(
                       alignment: Alignment.bottomCenter,
-                      key: _keyheight,
+                      //key: _keyheight,
                       child: ClipPath(
                           key: _keyheight,
                           clipper: NavBarClipper(currentIndex),
@@ -292,97 +371,160 @@ class _MainPageState extends State<MainPage> {
                               onClicked: onClicked,
                               selectedIndex: currentIndex,
                             ),
-                          ))),
-                ),
-              ],
-            ),
-          )
+                          )))
+                  // : Container(),
+                ],
+              ),
+            )),
+          ),
         ],
       );
     });
   }
 
-  // Stack buildFlexibleSpace(int currentIndex) {
-  //   return Stack(
-  //     children: <Widget>[
-  //       new SvgPicture.asset(
-  //         "assets/images/background_webreader.svg",
-  //         fit: BoxFit.fill,
-  //         allowDrawingOutsideViewBox: true,
-  //       ),
-  //       Scaffold(
-  //           extendBodyBehindAppBar: true,
-  //           // extendBody: true,
-  //           backgroundColor: Colors.transparent,
-  //           body: Stack(
-  //             children: [
-  //               ClipPath(
-  //                 clipper: BodyPainter(currentIndex, context, _keyheight),
-  //                 // clipBehavior: Clip.hardEdge,
-  //                 child: Container(
-  //                   child: BlocBuilder<NavbarBloc, NavbarState>(
-  //                       builder: (context, state) {
-  //                     if (state is GoToHome || state is GoToMenu) {
-  //                       return LocationAppbar();
-  //                       // return HomePageState();
-  //                       // } else if (state is GoToMenu) {
-  //                       // return MenuPage();
-  //                       return Container();
-  //                     } else if (state is GoToMap) {
-  //                       return MapsBlocWidget();
-  //                     } else if (state is GoToAccount) {
-  //                       return Account();
-  //                     }
-  //                     ;
-  //                     return Container();
-  //                   }),
-  //                 ),
-  //               ),
-  //               ClipPath(
-  //                 key: _keyheight,
-  //                 clipper: NavBarClipper(currentIndex),
-  //                 child: Container(
-  //                     alignment: Alignment.bottomCenter,
-  //                     key: _keyheight,
-  //                     child: ClipPath(
-  //                         key: _keyheight,
-  //                         clipper: NavBarClipper(currentIndex),
-  //                         child: CustomPaint(
-  //                           painter: NavBarPainter(currentIndex),
-  //                           child: BottomNavBarImp(
-  //                             onClicked: onClicked,
-  //                             selectedIndex: currentIndex,
-  //                           ),
-  //                         ))),
-  //               ),
-  //             ],
-  //           ))
-  //     ],
-  //   );
-  // }
+// Stack buildFlexibleSpace(int currentIndex) {
+//   return Stack(
+//     children: <Widget>[
+//       new SvgPicture.asset(
+//         "assets/images/background_webreader.svg",
+//         fit: BoxFit.fill,
+//         allowDrawingOutsideViewBox: true,
+//       ),
+//       Scaffold(
+//           extendBodyBehindAppBar: true,
+//           // extendBody: true,
+//           backgroundColor: Colors.transparent,
+//           body: Stack(
+//             children: [
+//               ClipPath(
+//                 clipper: BodyPainter(currentIndex, context, _keyheight),
+//                 // clipBehavior: Clip.hardEdge,
+//                 child: Container(
+//                   child: BlocBuilder<NavbarBloc, NavbarState>(
+//                       builder: (context, state) {
+//                     if (state is GoToHome || state is GoToMenu) {
+//                       return LocationAppbar();
+//                       // return HomePageState();
+//                       // } else if (state is GoToMenu) {
+//                       // return MenuPage();
+//                       return Container();
+//                     } else if (state is GoToMap) {
+//                       return MapsBlocWidget();
+//                     } else if (state is GoToAccount) {
+//                       return Account();
+//                     }
+//                     ;
+//                     return Container();
+//                   }),
+//                 ),
+//               ),
+//               ClipPath(
+//                 key: _keyheight,
+//                 clipper: NavBarClipper(currentIndex),
+//                 child: Container(
+//                     alignment: Alignment.bottomCenter,
+//                     key: _keyheight,
+//                     child: ClipPath(
+//                         key: _keyheight,
+//                         clipper: NavBarClipper(currentIndex),
+//                         child: CustomPaint(
+//                           painter: NavBarPainter(currentIndex),
+//                           child: BottomNavBarImp(
+//                             onClicked: onClicked,
+//                             selectedIndex: currentIndex,
+//                           ),
+//                         ))),
+//               ),
+//             ],
+//           ))
+//     ],
+//   );
+// }
 
-  // Stack buildLocationFlexibleSpace(int currentIndex) {
-  //   return Stack(
-  //     children: <Widget>[
-  //       new SvgPicture.asset(
-  //         "assets/images/background_webreader.svg",
-  //         fit: BoxFit.fill,
-  //         allowDrawingOutsideViewBox: true,
-  //       ),
-  //       Scaffold(
-  //           extendBodyBehindAppBar: true,
-  //           // extendBody: true,
-  //           backgroundColor: Colors.transparent,
-  //           body: Stack(
-  //             children: [
-  //               Container(
-  //                 alignment: Alignment.bottomCenter,
-  //                 // key: _keyheight,
-  //                 child: Text("data"),
-  //               )
-  //             ],
-  //           ))
-  //     ],
-  //   );
-  // }
+// Stack buildLocationFlexibleSpace(int currentIndex) {
+//   return Stack(
+//     children: <Widget>[
+//       new SvgPicture.asset(
+//         "assets/images/background_webreader.svg",
+//         fit: BoxFit.fill,
+//         allowDrawingOutsideViewBox: true,
+//       ),
+//       Scaffold(
+//           extendBodyBehindAppBar: true,
+//           // extendBody: true,
+//           backgroundColor: Colors.transparent,
+//           body: Stack(
+//             children: [
+//               Container(
+//                 alignment: Alignment.bottomCenter,
+//                 // key: _keyheight,
+//                 child: Text("data"),
+//               )
+//             ],
+//           ))
+//     ],
+//   );
+// }
 }
+
+// child: ExpansionTile(
+// // onExpansionChanged: (bool expanding) =>
+// //     _onExpansion(expanding),
+// leading: Icon(
+// Icons.circle,
+// color: Colors.white,
+// size: 80,
+// ),
+// trailing: GestureDetector(
+// // onTap: ,
+// child: Icon(
+// Icons.search,
+// color: Colors.white,
+// size: 30,
+// ),
+// ),
+// title: Row(
+// children: [
+// Column(
+// crossAxisAlignment: CrossAxisAlignment.start,
+// mainAxisAlignment: MainAxisAlignment.center,
+// children: [
+// Container(
+// child: Text(
+// "asd",
+// // state.magazinePublishedGetLastWithLimit
+// //     .response!.first.name!,
+// style: TextStyle(
+// fontSize: 18, color: Colors.white),
+// // textAlign: TextAlign.left,
+// ),
+// ),
+// Container(
+// child: Text(
+// "Infos zu deiner Location",
+// style: TextStyle(
+// fontSize: 13,
+// color: Colors.white,
+// fontWeight: FontWeight.w200),
+// // textAlign: TextAlign.right,
+// ),
+// ),
+// ],
+// ),
+// ],
+// ),
+// children: <Widget>[
+// SingleChildScrollView(
+// child: Container(
+// height: MediaQuery.of(context).size.height * 0.8,
+// width: MediaQuery.of(context).size.width,
+// child: Text(
+// "data",
+// style: TextStyle(
+// fontSize: 20, color: Colors.white),
+// textAlign: TextAlign.center,
+// ),
+// ),
+// ),
+// ],
+// ),
